requirements.txt
# Kafka Client
confluent-kafka==2.3.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
imbalanced-learn==0.11.0
xgboost==2.0.3

# Database
pyodbc==5.0.1

# Configuration
python-dotenv==1.0.0

# Model Persistence
joblib==1.3.2

# Logging
colorlog==6.8.0
